:white_check_mark: Add WSI processing for `AnnotationStore` output.
