+++ tests/engines/test_patch_predictor.py
@@ -3,9 +3,7 @@
 from __future__ import annotations
 
 import copy
-import json
 import shutil
-import sqlite3
 from pathlib import Path
 from typing import Callable
 
@@ -237,145 +235,6 @@ def test_wsi_predictor_api(
     assert accuracy > 0.99, np.nonzero(~diff)
 
     shutil.rmtree(_kwargs["save_dir"], ignore_errors=True)
-
-
-def _extract_probabilities_from_annotation_store(dbfile: str) -> dict:
-    """Helper function to extract probabilities from Annotation Store."""
-    probs_dict = {}
-    con = sqlite3.connect(dbfile)
-    cur = con.cursor()
-    annotations_properties = list(cur.execute("SELECT properties FROM annotations"))
-
-    for item in annotations_properties:
-        for json_str in item:
-            probs_dict = json.loads(json_str)
-            probs_dict.pop("prob_0")
-
-    return probs_dict
-
-
-def _validate_probabilities(predictions: list | dict) -> bool:
-    """Helper function to test if the probabilities value are valid."""
-    if isinstance(predictions, dict):
-        return all(0 <= probability <= 1 for _, probability in predictions.items())
-
-    for row in predictions:
-        for element in row:
-            if not (0 <= element <= 1):
-                return False
-    return True
-
-
-def test_wsi_predictor_zarr(sample_wsi_dict: dict, tmp_path: Path) -> None:
-    """Test normal run of patch predictor for WSIs."""
-    mini_wsi_svs = Path(sample_wsi_dict["wsi2_4k_4k_svs"])
-
-    classifier = PatchPredictor(
-        model="alexnet-kather100k",
-        batch_size=32,
-        verbose=False,
-    )
-    # don't run test on GPU
-    output = classifier.run(
-        images=[mini_wsi_svs],
-        return_probabilities=True,
-        return_labels=False,
-        device=device,
-        patch_mode=False,
-        save_dir=tmp_path / "wsi_out_check",
-    )
-
-    assert output[mini_wsi_svs].exists()
-
-    output_ = zarr.open(output[mini_wsi_svs])
-
-    assert output_["probabilities"].shape == (70, 9)  # number of patches x classes
-    assert output_["probabilities"].ndim == 2
-    # number of patches x [start_x, start_y, end_x, end_y]
-    assert output_["coordinates"].shape == (70, 4)
-    assert output_["coordinates"].ndim == 2
-    assert _validate_probabilities(predictions=output_["probabilities"])
-
-
-def test_engine_run_wsi_annotation_store(
-    sample_wsi_dict: dict,
-    tmp_path: Path,
-) -> None:
-    """Test the engine run for Whole slide images."""
-    # convert to pathlib Path to prevent wsireader complaint
-    mini_wsi_svs = Path(sample_wsi_dict["wsi2_4k_4k_svs"])
-    mini_wsi_msk = Path(sample_wsi_dict["wsi2_4k_4k_msk"])
-
-    eng = PatchPredictor(model="alexnet-kather100k")
-
-    patch_size = np.array([224, 224])
-    save_dir = f"{tmp_path}/model_wsi_output"
-
-    kwargs = {
-        "patch_input_shape": patch_size,
-        "stride_shape": patch_size,
-        "resolution": 0.5,
-        "save_dir": save_dir,
-        "units": "mpp",
-        "scale_factor": (2.0, 2.0),
-    }
-
-    output = eng.run(
-        images=[mini_wsi_svs],
-        masks=[mini_wsi_msk],
-        patch_mode=False,
-        output_type="AnnotationStore",
-        **kwargs,
-    )
-
-    output_ = output[mini_wsi_svs]
-
-    assert output_.exists()
-    assert output_.suffix == ".db"
-    predictions = _extract_probabilities_from_annotation_store(output_)
-    assert _validate_probabilities(predictions)
-
-    shutil.rmtree(save_dir)
-
-
-def test_engine_run_wsi_annotation_store_power(
-    sample_wsi_dict: dict,
-    tmp_path: Path,
-) -> None:
-    """Test the engine run for Whole slide images."""
-    # convert to pathlib Path to prevent wsireader complaint
-    mini_wsi_svs = Path(sample_wsi_dict["wsi2_4k_4k_svs"])
-    mini_wsi_msk = Path(sample_wsi_dict["wsi2_4k_4k_msk"])
-
-    eng = PatchPredictor(model="alexnet-kather100k")
-
-    patch_size = np.array([224, 224])
-    save_dir = f"{tmp_path}/model_wsi_output"
-
-    kwargs = {
-        "patch_input_shape": patch_size,
-        "stride_shape": patch_size,
-        "resolution": 20,
-        "save_dir": save_dir,
-        "units": "power",
-    }
-
-    output = eng.run(
-        images=[mini_wsi_svs],
-        masks=[mini_wsi_msk],
-        patch_mode=False,
-        output_type="AnnotationStore",
-        **kwargs,
-    )
-
-    output_ = output[mini_wsi_svs]
-
-    assert output_.exists()
-    assert output_.suffix == ".db"
-    predictions = _extract_probabilities_from_annotation_store(output_)
-    assert _validate_probabilities(predictions)
-
-    shutil.rmtree(save_dir)
 
 
 # -------------------------------------------------------------------------------------
